requirements.txt
bidict==0.23.1
blinker==1.8.2
bokeh==3.6.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.3.0
dash==2.18.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dnspython==2.6.1
eventlet==0.37.0
Flask==2.3.2
Flask-Cors==5.0.0
Flask-SocketIO==5.3.7
future==1.0.0
greenlet==3.1.1
h11==0.14.0
idna==3.8
importlib_metadata==8.5.0
iso8601==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
nest-asyncio==1.6.0
numpy==1.26.0
packaging==24.1
pandas==2.2.3
pika==1.3.2
pillow==11.0.0
plotly==5.24.1
pyserial==3.5
python-dateutil==2.9.0.post0
python-engineio==4.9.1
python-socketio==5.11.4
pytz==2024.2
PyYAML==6.0.2
requests==2.31.0
retrying==1.3.4
setuptools==75.1.0
simple-websocket==1.0.0
six==1.16.0
tenacity==9.0.0
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.2
Werkzeug==3.0.4
wheel==0.44.0
wsproto==1.2.0
xyzservices==2024.9.0
zipp==3.20.2
